[shared] Move Timeout into utils
