+++ backend/runner/src/main/java/de/aaaaaaah/velcom/runner/Connection.java
@@ -2,7 +2,7 @@ package de.aaaaaaah.velcom.runner;
 
 import de.aaaaaaah.velcom.runner.states.Idle;
 import de.aaaaaaah.velcom.runner.states.RunnerState;
-import de.aaaaaaah.velcom.shared.Timeout;
+import de.aaaaaaah.velcom.shared.util.Timeout;
 import de.aaaaaaah.velcom.shared.protocol.RunnerConnectionHeader;
 import de.aaaaaaah.velcom.shared.protocol.StatusCode;
 import de.aaaaaaah.velcom.shared.protocol.serialization.Serializer;
